Move relevent functions to surface_analysis.py
