+++ src/Func_pyrough.py
@@ -1823,66 +1823,3 @@ def center_3d_dataset(dataset):
     )
 
     return centered_dataset
-
-
-####################################################################
-####################### SURFACE ANALYSIS ###########################
-####################################################################
-
-
-def rough(x, y, eta, RMS, M, N):
-    """
-    Generates a rough surface
-
-    :param x: List of x coordinates
-    :type x: array
-    :param y: List of y coordinates
-    :type y: array
-    :param eta: Roughness exponent
-    :type eta: float
-    :param C1: Normalization factor
-    :type C1: float
-    :param M: Scaling cartesian position
-    :type M: int
-    :param N: Scaling cartesian position
-    :type N: int
-
-    :return: Height matrix
-    """
-    z = np.zeros(np.shape(x))
-    listM = np.linspace(-M, M, 2 * M + 1)
-    listN = np.linspace(-N, N, 2 * N + 1)
-    k = 0
-    for m in listM:
-        for n in listN:
-            if m == 0 and n == 0:
-                continue
-            else:
-                mod = (m * m + n * n) ** (-1 * (1 + eta))
-                G = np.random.randn()
-                U = -np.pi / 2 + np.pi * np.random.rand()
-                zadd = G * mod * np.cos(2 * np.pi * (m * x + n * y) + U)
-                z = z + zadd
-            k += 1
-    RMS_i = rms_calc(z)
-    C1 = RMS / RMS_i
-    return C1 * z
-
-
-def rescale(D, scale):
-    """
-    Rescales the height distribution between zmin and zmax
-
-    :param D: Height matrix
-    :type D: array
-    :param scale: zmin and zmax
-    :type scale: list
-
-    :return: Rescaled matrix
-    """
-    D = D - np.min(D)
-    D = D / np.max(D)
-    lower = scale[0]
-    upper = scale[1]
-    Df = [lower + (upper - lower) * x for x in D]
-    return np.asarray(Df)
